Improve addition of AU regions
This addresses issues added in later versions of Magento where the
Australia regions were added to the core product. It should also improve
situations where other extensions may have added the regions already as
it will simply check if the regions exist in the database and not add
them if so.